refactor: move media/net from brightray to atom
