+++ atom/browser/media/media_stream_devices_controller.cc
@@ -2,16 +2,16 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE-CHROMIUM file.
 
-#include "brightray/browser/media/media_stream_devices_controller.h"
+#include "atom/browser/media/media_stream_devices_controller.h"
 
 #include <memory>
 #include <utility>
 
-#include "brightray/browser/media/media_capture_devices_dispatcher.h"
+#include "atom/browser/media/media_capture_devices_dispatcher.h"
 #include "content/public/browser/desktop_media_id.h"
 #include "content/public/common/media_stream_request.h"
 
-namespace brightray {
+namespace atom {
 
 namespace {
 
@@ -188,4 +188,4 @@ void MediaStreamDevicesController::HandleUserMediaRequest() {
                            std::unique_ptr<content::MediaStreamUI>());
 }
 
-}  // namespace brightray
+}  // namespace atom
