+++ atom/browser/net/atom_cert_verifier.h
@@ -11,15 +11,10 @@
 
 #include "net/cert/cert_verifier.h"
 
-namespace brightray {
-
-class RequireCTDelegate;
-
-}  // namespace brightray
-
 namespace atom {
 
 class CertVerifierRequest;
+class RequireCTDelegate;
 
 struct VerifyRequestParams {
   std::string hostname;
@@ -34,7 +29,7 @@ struct VerifyRequestParams {
 
 class AtomCertVerifier : public net::CertVerifier {
  public:
-  explicit AtomCertVerifier(brightray::RequireCTDelegate* ct_delegate);
+  explicit AtomCertVerifier(RequireCTDelegate* ct_delegate);
   ~AtomCertVerifier() override;
 
   using VerifyProc = base::Callback<void(const VerifyRequestParams& request,
@@ -43,7 +38,7 @@ class AtomCertVerifier : public net::CertVerifier {
   void SetVerifyProc(const VerifyProc& proc);
 
   const VerifyProc verify_proc() const { return verify_proc_; }
-  brightray::RequireCTDelegate* ct_delegate() const { return ct_delegate_; }
+  RequireCTDelegate* ct_delegate() const { return ct_delegate_; }
   net::CertVerifier* default_verifier() const {
     return default_cert_verifier_.get();
   }
@@ -66,7 +61,7 @@ class AtomCertVerifier : public net::CertVerifier {
   std::map<RequestParams, CertVerifierRequest*> inflight_requests_;
   VerifyProc verify_proc_;
   std::unique_ptr<net::CertVerifier> default_cert_verifier_;
-  brightray::RequireCTDelegate* ct_delegate_;
+  RequireCTDelegate* ct_delegate_;
 
   DISALLOW_COPY_AND_ASSIGN(AtomCertVerifier);
 };
