+++ atom/browser/net/require_ct_delegate.cc
@@ -2,11 +2,11 @@
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#include "brightray/browser/net/require_ct_delegate.h"
+#include "atom/browser/net/require_ct_delegate.h"
 
 #include "content/public/browser/browser_thread.h"
 
-namespace brightray {
+namespace atom {
 
 RequireCTDelegate::RequireCTDelegate() {}
 
@@ -33,4 +33,4 @@ RequireCTDelegate::CTRequirementLevel RequireCTDelegate::IsCTRequiredForHost(
   return CTRequirementLevel::DEFAULT;
 }
 
-}  // namespace brightray
+}  // namespace atom
