+++ brightray/BUILD.gn
@@ -24,18 +24,8 @@ static_library("brightray") {
     "browser/browser_main_parts.cc",
     "browser/browser_main_parts.h",
     "browser/browser_main_parts_mac.mm",
-    "browser/media/media_capture_devices_dispatcher.cc",
-    "browser/media/media_capture_devices_dispatcher.h",
-    "browser/media/media_device_id_salt.cc",
-    "browser/media/media_device_id_salt.h",
-    "browser/media/media_stream_devices_controller.cc",
-    "browser/media/media_stream_devices_controller.h",
-    "browser/net/require_ct_delegate.cc",
-    "browser/net/require_ct_delegate.h",
     "browser/win/scoped_hstring.cc",
     "browser/win/scoped_hstring.h",
-    "browser/zoom_level_delegate.cc",
-    "browser/zoom_level_delegate.h",
     "common/application_info.cc",
     "common/application_info.h",
     "common/application_info_mac.mm",
@@ -47,7 +37,5 @@ static_library("brightray") {
     "common/main_delegate.cc",
     "common/main_delegate.h",
     "common/main_delegate_mac.mm",
-    "common/platform_util.h",
-    "common/platform_util_linux.cc",
   ]
 }
