+++ filenames.gni
@@ -288,6 +288,12 @@ filenames = {
     "atom/browser/native_window_mac.h",
     "atom/browser/native_window_mac.mm",
     "atom/browser/native_window_observer.h",
+    "atom/browser/media/media_capture_devices_dispatcher.cc",
+    "atom/browser/media/media_capture_devices_dispatcher.h",
+    "atom/browser/media/media_device_id_salt.cc",
+    "atom/browser/media/media_device_id_salt.h",
+    "atom/browser/media/media_stream_devices_controller.cc",
+    "atom/browser/media/media_stream_devices_controller.h",
     "atom/browser/net/about_protocol_handler.cc",
     "atom/browser/net/about_protocol_handler.h",
     "atom/browser/net/asar/asar_protocol_handler.cc",
@@ -306,6 +312,8 @@ filenames = {
     "atom/browser/net/http_protocol_handler.h",
     "atom/browser/net/js_asker.cc",
     "atom/browser/net/js_asker.h",
+    "atom/browser/net/require_ct_delegate.cc",
+    "atom/browser/net/require_ct_delegate.h",
     "atom/browser/net/resolve_proxy_helper.cc",
     "atom/browser/net/resolve_proxy_helper.h",
     "atom/browser/net/url_request_about_job.cc",
@@ -503,6 +511,8 @@ filenames = {
     "atom/browser/window_list.cc",
     "atom/browser/window_list.h",
     "atom/browser/window_list_observer.h",
+    "atom/browser/zoom_level_delegate.cc",
+    "atom/browser/zoom_level_delegate.h",
     "atom/common/api/api_messages.h",
     "atom/common/api/atom_api_asar.cc",
     "atom/common/api/atom_api_clipboard.cc",
